refactor: move interfaces to entities folders
