+++ src/album/entities/album.model.ts
@@ -1,0 +1,6 @@
+export interface IAlbum {
+  id: string; // uuid v4
+  name: string;
+  year: number;
+  artistId: string | null; // refers to Artist
+}
