+++ src/artist/entities/artist.model.ts
@@ -1,0 +1,5 @@
+export interface IArtist {
+  id: string; // uuid v4
+  name: string;
+  grammy: boolean;
+}
