--- src/interfaces.ts
@@ -1,45 +1,0 @@
-import { Album } from './album/entities/album.entity';
-import { Artist } from './artist/entities/artist.entity';
-import { Track } from './track/entities/track.entity';
-
-export interface IUser {
-  id: string; // uuid v4
-  login: string;
-  password: string;
-  version: number; // integer number, increments on update
-  createdAt: number; // timestamp of creation
-  updatedAt: number; // timestamp of last update
-}
-
-export interface IArtist {
-  id: string; // uuid v4
-  name: string;
-  grammy: boolean;
-}
-
-export interface ITrack {
-  id: string; // uuid v4
-  name: string;
-  artistId: string | null; // refers to Artist
-  albumId: string | null; // refers to Album
-  duration: number; // integer number
-}
-
-export interface IAlbum {
-  id: string; // uuid v4
-  name: string;
-  year: number;
-  artistId: string | null; // refers to Artist
-}
-
-export interface IFavorites {
-  artists: string[]; // favorite artists ids
-  albums: string[]; // favorite albums ids
-  tracks: string[]; // favorite tracks ids
-}
-
-export interface IFavoritesResponse {
-  artists: Artist[];
-  albums: Album[];
-  tracks: Track[];
-}
