+++ src/user/entities/user.model.ts
@@ -1,0 +1,8 @@
+export interface IUser {
+  id: string; // uuid v4
+  login: string;
+  password: string;
+  version: number; // integer number, increments on update
+  createdAt: number; // timestamp of creation
+  updatedAt: number; // timestamp of last update
+}
